requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
openai==1.51.0
requests==2.32.3
